backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.0.1
motor==3.3.1
pymongo==4.5.0
pydantic==2.6.4
pydantic-settings==2.2.1
bcrypt==4.1.2
PyJWT==2.8.0
aiofiles==23.2.1
python-multipart==0.0.9
email-validator==2.1.1
httpx==0.27.0
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-analytics-data
